game service created, calling service in add screen
